- реализовано взаимодействие с полосой загрузки - реализована остановка скачивания книги - улучшен переход в меню поиска книги

+++ ABPlayer/ui_functions/book_page.py
@@ -8,14 +8,16 @@ import urllib.request
 from PyQt5.QtCore import (
     Qt,
     QSize,
+    QEvent,
     QThread,
     QObject,
     pyqtSignal,
+    QBasicTimer,
     QEasingCurve,
     QPropertyAnimation,
 )
 from PyQt5.QtGui import QMovie, QPixmap
-from PyQt5.QtWidgets import QMessageBox
+from PyQt5.QtWidgets import QMessageBox, QToolTip, QProgressBar
 
 from database import Books
 from drivers import drivers, BaseDownloadProcessHandler
@@ -25,6 +27,16 @@ if ty.TYPE_CHECKING:
     from PyQt5.QtWidgets import QLabel
     from main_window import MainWindow
     from database import Book
+
+
+def convert_into_bits(bits: int) -> str:
+    postfixes = ["КБ", "МБ", "ГБ"]
+    if bits >= 2 ** 33:
+        return f"{round(bits / 2 ** 33, 3)} {postfixes[-1]}"
+    elif bits >= 2 ** 23:
+        return f"{round(bits / 2 ** 23, 2)} {postfixes[-2]}"
+    elif bits >= 2 ** 13:
+        return f"{round(bits / 2 ** 13, 1)} {postfixes[-3]}"
 
 
 class Cache(object):
@@ -157,19 +169,21 @@ class DownloadBookWorker(QObject):
 
     finished: QtCore.pyqtSignal = pyqtSignal()
     failed: QtCore.pyqtSignal = pyqtSignal(str)
+    close: QtCore.pyqtSignal = pyqtSignal()
 
     def __init__(self, main_window: MainWindow, book: Book):
         super(DownloadBookWorker, self).__init__()
         self.main_window, self.book = main_window, book
+        self.drv = [drv for drv in drivers if self.book.url.startswith(drv().site_url)][
+            0
+        ]()  # Драйвер, который нужно использовать для скачивания
         self.finished.connect(self.finish)
         self.failed.connect(lambda text: self.fail(text))
+        self.close.connect(self._close)
 
     def run(self):
         try:
-            drv = [drv for drv in drivers if self.book.url.startswith(drv().site_url)][
-                0
-            ]()  # Драйвер, который нужно использовать для скачивания
-            drv.download_book(self.book, DownloadProcessHandler(self.main_window))
+            self.drv.download_book(self.book, DownloadProcessHandler(self.main_window))
             books = Books(os.environ["DB_PATH"])
             books.insert(**vars(self.book))  # Добавляем книгу в бд
             self.finished.emit()
@@ -194,6 +208,11 @@ class DownloadBookWorker(QObject):
             self.main_window.pb_animation.setEndValue(0)
             self.main_window.pb_animation.setEasingCurve(QEasingCurve.InOutQuart)
             self.main_window.pb_animation.start()
+            if (
+                self.main_window.stackedWidget.currentWidget()
+                == self.main_window.libraryPage
+            ):
+                self.main_window.openLibraryPage()
 
     def fail(self, text: str):
         self.main_window.download_book_thread.quit()
@@ -205,6 +224,11 @@ class DownloadBookWorker(QObject):
             ),
         )
 
+    def _close(self):
+        file = self.drv.__dict__.get("_file")
+        if file:
+            file.close()
+
 
 class DownloadProcessHandler(BaseDownloadProcessHandler):
     def __init__(self, main_window: MainWindow):
@@ -213,6 +237,7 @@ class DownloadProcessHandler(BaseDownloadProcessHandler):
         """
         super(DownloadProcessHandler, self).__init__()
         self.main_window = main_window
+        self._last_size = ""
 
     def show_progress(self) -> None:
         """
@@ -220,6 +245,33 @@ class DownloadProcessHandler(BaseDownloadProcessHandler):
         """
         progress = int(round(self.done_size / (self.total_size / 100), 0))
         self.main_window.downloadingProgressBarLg.setValue(progress)
+        self.main_window.downloadingProgressBar.setValue(progress)
+        self.main_window.downloadingProgressBar.setToolTip(
+            f"{convert_into_bits(self.done_size)} / {convert_into_bits(self.total_size)}",
+        )
+
+
+def prepareProgressBar(pb: QProgressBar) -> None:
+    class ToolTipUpdater(QObject):
+        def timerEvent(self, x) -> None:
+            QToolTip.showText(pb.toolTipPos, pb.toolTip(), pb)
+
+    def eventFilter(event: QEvent) -> bool:
+        if event.type() == QEvent.ToolTip:
+            pb.toolTipPos = event.globalPos()
+            pb.toolTipUpdater = ToolTipUpdater()
+            pb.toolTipTimer = QBasicTimer()
+            pb.toolTipTimer.start(100, pb.toolTipUpdater)
+            return True
+        elif event.type() == QEvent.Leave:
+            if pb.__dict__.get("toolTipPos"):
+                pb.toolTipTimer.stop()
+                pb.toolTipPos = None
+                QToolTip.hideText()
+        return QProgressBar.event(pb, event)
+
+    pb.toolTipPos = None
+    pb.event = eventFilter
 
 
 def download_book(main_window: MainWindow, book: Book) -> None:
@@ -253,6 +305,36 @@ def download_book(main_window: MainWindow, book: Book) -> None:
     main_window.download_book_thread.start()
 
 
+def stopBookDownloading(main_window: MainWindow) -> None:
+    answer = QMessageBox.question(
+        main_window,
+        "Подтвердите действие",
+        "Вы действительно хотите прервать скачивание книги?",
+        QMessageBox.Yes | QMessageBox.No,
+        QMessageBox.No,
+    )
+
+    if answer == QMessageBox.No:
+        return
+
+    main_window.download_book_worker.close.emit()
+    main_window.download_book_thread.terminate()
+    downloadable_book = main_window.downloadable_book
+    main_window.downloadable_book = ...
+
+    # Открываем страницу загрузки
+    main_window.delete_book_loading_movie = QMovie(":/other/loading.gif")
+    main_window.delete_book_loading_movie.setScaledSize(QSize(50, 50))
+    main_window.openInfoPage(movie=main_window.delete_book_loading_movie)
+
+    # Создаем и запускаем новый поток
+    main_window.delete_book_thread = QThread()
+    main_window.delete_book_worker = DeleteBookWorker(main_window, downloadable_book)
+    main_window.delete_book_worker.moveToThread(main_window.delete_book_thread)
+    main_window.delete_book_thread.started.connect(main_window.delete_book_worker.run)
+    main_window.delete_book_thread.start()
+
+
 class DeleteBookWorker(QObject):
     """
     Реализует удаление книги.
@@ -261,7 +343,7 @@ class DeleteBookWorker(QObject):
     finished: QtCore.pyqtSignal = pyqtSignal()
     failed: QtCore.pyqtSignal = pyqtSignal(str)
 
-    def __init__(self, main_window: MainWindow, book: Books):
+    def __init__(self, main_window: MainWindow, book: ty.Union[Book, Books]):
         super(DeleteBookWorker, self).__init__()
         self.main_window, self.book = main_window, book
         self.finished.connect(self.finish)
@@ -271,13 +353,15 @@ class DeleteBookWorker(QObject):
         try:
             self.main_window.btnGroupFrame.setDisabled(True)
             self.main_window.btnGroupFrame_2.setDisabled(True)
-            books = Books(os.environ["DB_PATH"])
-            books.api.execute("""DELETE FROM books WHERE id=?""", self.book.id)
-            books.api.commit()
-            for root, dirs, files in os.walk(self.book.dir_path):
-                for file in files:
-                    os.remove(os.path.join(root, file))
-                os.rmdir(root)
+            if self.book.__dict__.get("id"):
+                books = Books(os.environ["DB_PATH"])
+                books.api.execute("""DELETE FROM books WHERE id=?""", self.book.id)
+                books.api.commit()
+            if os.path.isdir(self.book.dir_path):
+                for root, dirs, files in os.walk(self.book.dir_path):
+                    for file in files:
+                        os.remove(os.path.join(root, file))
+                    os.rmdir(root)
             self.finished.emit()
         except Exception as err:
             # TODO: Необходимо реализовать нормальный обзор ошибок
